Move java.io.Console operations to jdk.internal.io.JdkConsoleImpl
This is part of a simplified backport of these two changes:

8298416: Console should be declared `sealed`
8298971: Move Console implementation into jdk internal package

This commit shows the Java code movement without noise (though the
result does not compile).

Note that the removal of the line in LineReader.read:

end = cbuf.length;

shortly after the "no space left..." comment, is intentional.  It was
approved in the original pull request:

https://github.com/openjdk/jdk/pull/11615#discussion_r1047942934

Copy src/java.base/share/classes/jdk/internal/io/JdkConsoleImpl.java
as of jdk21u-dev 24b3b2b66f0bed3e9812999a5b6d511f31e9ad1f.
